some updates

+++ sources/client.c
@@ -6,11 +6,31 @@
 /*   By: rojornod <rojornod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/25 10:33:05 by rojornod          #+#    #+#             */
-/*   Updated: 2025/02/27 17:00:22 by rojornod         ###   ########.fr       */
+/*   Updated: 2025/02/28 16:33:19 by rojornod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minitalk.h"
+
+int	mod_atoi(const char *str)
+{
+	int	i;
+	int	sign;
+	int	result;
+
+	i = 0;
+	result = 0;
+	sign = 1;
+	while (str[i] != '\0')
+	{
+		if (str[i] < '0' || str[i] > '9')
+			return (-1);
+		else 
+			result = result * 10 + (str[i] - 48);
+		i++;
+	}
+	return (result * sign);
+}
 
 static void	send_signal(char *message, int pid)
 {
@@ -26,7 +46,6 @@ static void	send_signal(char *message, int pid)
 		{
 			
 			byte_index = message[i] >> (7-bit) & 1;
-			//ft_printf("%d", byte_index);
 			if (byte_index == 1)
 				kill(pid, SIGUSR1);
 			if (byte_index == 0)
@@ -36,6 +55,7 @@ static void	send_signal(char *message, int pid)
 		}
 		i++;
 	}
+	
 }
 
 int main(int argc, char **argv)
@@ -44,7 +64,12 @@ int main(int argc, char **argv)
 
 	if (argc != 3)
 		return (ft_printf("Error, try again\n"), EXIT_FAILURE);
-	pid = ft_atoi(argv[1]);
+	pid = mod_atoi(argv[1]);
+	if (pid == -1)
+	{
+		ft_printf("Invalid process ID\n");
+		return (-1);
+	}
 	send_signal(argv[2], pid);
 	return (0);
 }
